+++ src/plugins/blocksmith/components/eme-form-block-modal/index.jsx
@@ -1,7 +1,7 @@
 import DateTimePicker from '@triniti/cms/plugins/blocksmith/components/date-time-picker';
-import DatePickerField from '@triniti/cms/components/date-picker-field';
 import Message from '@gdbots/pbj/Message';
 import NodeRef from '@gdbots/schemas/gdbots/ncr/NodeRef';
+import moment from 'moment';
 import PropTypes from 'prop-types';
 import React from 'react';
 import {
@@ -12,11 +12,8 @@ import {
   Modal,
   ModalHeader,
   ModalBody,
-  ModalFooter,
+  ModalFooter, Checkbox,
 } from '@triniti/admin-ui-plugin/components';
-
-import changedDate from '../../utils/changedDate';
-import changedTime from '../../utils/changedTime';
 
 export default class EmeFormBlockModal extends React.Component {
   static propTypes = {
@@ -42,13 +39,14 @@ export default class EmeFormBlockModal extends React.Component {
       touched: false,
       isValid: block.has('form_ref'),
       formRef: block.get('form_ref'),
-      expiresDate: block.get('expires_at', new Date()),
+      hasExpireDate: block.has('expires_at'),
+      expireDate: block.get('expires_at', new Date()),
     };
 
     this.handleAddBlock = this.handleAddBlock.bind(this);
     this.handleChangeFormRef = this.handleChangeFormRef.bind(this);
     this.handleEditBlock = this.handleEditBlock.bind(this);
-
+    this.handleChangeCheckbox = this.handleChangeCheckbox.bind(this);
     this.handleChangeDate = this.handleChangeDate.bind(this);
     this.handleChangeTime = this.handleChangeTime.bind(this);
   }
@@ -91,41 +89,37 @@ export default class EmeFormBlockModal extends React.Component {
     });
   }
 
-  setBlock() {
-    const {
-      formRef, expiresDate
-    } = this.state;
+  handleChangeCheckbox({ target: { checked, id } }) {
+    this.setState({ [id]: checked });
+  }
 
+  handleChangeDate(date) {
+    this.setState({
+      expireDate: moment(date)
+        .set('year', date.getFullYear())
+        .set('month', date.getMonth())
+        .set('date', date.getDate())
+        .toDate(),
+    });
+  }
+
+  handleChangeTime({ target: { value: time } }) {
+    const { expireDate } = this.state;
+    this.setState({
+      expireDate: moment(expireDate)
+        .set('hours', time ? time.split(':')[0] : 0)
+        .set('minutes', time ? time.split(':')[1] : 0)
+        .toDate(),
+    });
+  }
+
+  setBlock() {
+    const { formRef, hasExpireDate, expireDate } = this.state;
     const { block } = this.props;
 
     return block.schema().createMessage()
       .set('form_ref', formRef || null)
-      .set('expires_at', expiresDate || null);
-  }
-
-  /*handleChangeDate(date) {
-    this.setState(changedDate(date));
-  }*/
-
-  handleChangeDate(key, value) {
-    const { request } = this.state;
-
-    if (value === '') {
-      return;
-    }
-
-    if (value === null || value instanceof Date) {
-      this.setState({
-        request: {
-          ...request,
-          [key]: value,
-        },
-      });
-    }
-  }
-
-  handleChangeTime({ target: { value: time } }) {
-    this.setState(changedTime(time));
+      .set('expires_at', hasExpireDate ? expireDate : null);
   }
 
   render() {
@@ -134,7 +128,8 @@ export default class EmeFormBlockModal extends React.Component {
       isValid,
       touched,
       formRef,
-      expiresDate
+      hasExpireDate,
+      expireDate,
     } = this.state;
     const { isFreshBlock, isOpen, toggle } = this.props;
 
@@ -152,21 +147,21 @@ export default class EmeFormBlockModal extends React.Component {
               value={formRef || ''}
             />
           </FormGroup>
-          {/*<DateTimePicker
-            onChangeDate={this.handleChangeDate}
-            onChangeTime={this.handleChangeTime}
-            updatedDate={expiresDate}
-          />*/}
-          <DatePickerField
-            input={{
-              name: 'Expires At',
-              onChange: (date) => this.handleChangeDate('expires_at', date),
-              value: expiresDate || null,
-            }}
-            label="Expires At"
-            meta={{ error: '' }}
-            showSetCurrentDateTimeIcon={true}
-          />
+          <FormGroup className="mr-4">
+            <Checkbox size="sd" id="hasExpireDate" checked={hasExpireDate} onChange={this.handleChangeCheckbox}>
+              Has expiration date
+            </Checkbox>
+          </FormGroup>
+          {hasExpireDate
+          && (
+            <div className="modal-body-blocksmith">
+              <DateTimePicker
+                onChangeDate={this.handleChangeDate}
+                onChangeTime={this.handleChangeTime}
+                updatedDate={expireDate}
+              />
+            </div>
+          )}
           {
             !isValid && touched && formRef !== ''
             && <p className="text-danger">{errorMsg}</p>